+++ src/peer/component/mod.rs
@@ -153,6 +153,9 @@ pub struct State {
 
     /// [`SyncPhase`] of this [`Component`].
     sync_phase: ObservableCell<SyncPhase>,
+
+    /// Interval (in milliseconds) of [`PeerConnection`]'s stats scraping.
+    stats_scrape_interval_ms: u32,
 }
 
 impl State {
@@ -164,6 +167,7 @@ impl State {
         force_relay: bool,
         negotiation_role: Option<NegotiationRole>,
         connection_mode: ConnectionMode,
+        stats_scrape_interval_ms: u32,
     ) -> Self {
         Self {
             id,
@@ -182,6 +186,7 @@ impl State {
             maybe_update_local_stream: ObservableCell::new(false),
             maybe_update_connections: ObservableCell::new(None),
             sync_phase: ObservableCell::new(SyncPhase::Synced),
+            stats_scrape_interval_ms,
         }
     }
 
@@ -195,6 +200,12 @@ impl State {
     #[must_use]
     pub const fn id(&self) -> Id {
         self.id
+    }
+
+    /// Returns the stats scraping interval of this [`State`] in milliseconds.
+    #[must_use]
+    pub const fn stats_scrape_interval_ms(&self) -> u32 {
+        self.stats_scrape_interval_ms
     }
 
     /// Returns all [`IceServer`]s of this [`State`].
@@ -481,6 +492,7 @@ impl AsProtoState for State {
             local_sdp: self.local_sdp.current(),
             remote_sdp: self.remote_sdp.get(),
             restart_ice: self.restart_ice.get(),
+            stats_scrape_interval_ms: self.stats_scrape_interval_ms,
         }
     }
 }
@@ -498,6 +510,7 @@ impl SynchronizableState for State {
             input.force_relay,
             input.negotiation_role,
             input.connection_mode,
+            input.stats_scrape_interval_ms,
         );
 
         #[expect(clippy::iter_over_hash_type, reason = "order doesn't matter")]