+++ src/room.rs
@@ -1559,6 +1559,7 @@ impl EventHandler for InnerRoom {
         tracks: Vec<Track>,
         ice_servers: Vec<IceServer>,
         force_relay: bool,
+        stats_scrape_interval_ms: u32,
     ) -> Self::Output {
         let peer_state = peer::State::new(
             peer_id,
@@ -1566,6 +1567,7 @@ impl EventHandler for InnerRoom {
             force_relay,
             Some(negotiation_role),
             connection_mode,
+            stats_scrape_interval_ms,
         );
         for track in &tracks {
             peer_state.insert_track(track, self.send_constraints.clone());